GH-14: Move Resources to seperate file
Assosiated schemas were moved as well, I think that it will be
more maintainable to keep schemas and resources together (I could
easily be wrong about that though).
